Add support for _bulk links.
Supported fields:
*  `pk`
*  `filters`

Usage: for link to answer of result use `<0>`,
where 0 is index of requests and answers. If you need to get data from
result, use it - `<0[data][id]>`,
where 0 is index of requests and answers, 'data' is answer field with data
and 'id' is field in request data answer.

Close: vst/vst-utils#9